Revert "ci(convex): Scan the full JavaScript identifier grammar"
This reverts commit 7d8f35d4cc1351252b7ac258b9d3e2307bf5a18d.
